requirements.txt
gunicorn
Flask
Werkzeug
pydantic
azure-identity
pyodbc
python-dotenv
flask_jwt_extended
azure.storage.blob